Copy over HTML+CSS+JS for real-time dashboard
... from where it used to live in
https://github.com/Khan/analytics/blob/master/webapps/dashboards

Will remove from that repo once I get this one working.

Don't have to review this since this is just a copy. Will base future changes on
this so the diffs are just new/changed code.

Test Plan:
Not tested! This is just a copy of a bunch of files and does not actually work.
This is just so future commits can be diffed against this original.

Auditors: benkomalo, alpert