Design intervention.
People were asking me about the font on this thing. Seriously, there
were inquiries. I tried to distance myself from it but I felt something
had to be done.

I just want it to be noted that David must have felt pretty strongly
about including that font in the first place. He had to go to Google's
font directory looking for the one that would complete his dashboard.
The out of all the fonts they had available he had to say "That one.
That's the font I need. There is no better font for my dashboard."

Seriously, it boggles the mind.

Auditors: david